+++ src/main-thread/configuration.ts
@@ -17,21 +17,8 @@
 import { MessageFromWorker, MessageToWorker } from '../transfer/Messages';
 import { Phase } from '../transfer/Phase';
 import { HydrateableNode } from '../transfer/TransferrableNodes';
-import { DefaultAllowedMutations, TransferrableMutationType } from '../transfer/TransferrableMutation';
+import { DefaultAllowedMutations } from '../transfer/TransferrableMutation';
 import { CommandExecutorInterface } from './commands/interface';
-import { LongTaskExecutor } from './commands/long-task';
-import { ChildListProcessor } from './commands/child-list';
-import { AttributeProcessor } from './commands/attribute';
-import { CharacterDataProcessor } from './commands/character-data';
-import { PropertyProcessor } from './commands/property';
-import { EventSubscriptionProcessor } from './commands/event-subscription';
-import { BoundingClientRectProcessor } from './commands/bounding-client-rect';
-import { OffscreenCanvasProcessor } from './commands/offscreen-canvas';
-import { ObjectMutationProcessor } from './commands/object-mutation';
-import { ObjectCreationProcessor } from './commands/object-creation';
-import { ImageBitmapProcessor } from './commands/image-bitmap';
-import { StorageProcessor } from './commands/storage';
-import { FunctionProcessor } from './commands/function';
 
 /**
  * The callback for `mutationPump`. If specified, this callback will be called
@@ -112,30 +99,4 @@ export function normalizeConfiguration(
     },
     config,
   );
-}
-
-export function getLiteProcessors() {
-  return {
-    [TransferrableMutationType.FUNCTION_CALL]: FunctionProcessor,
-  };
-}
-
-export function getAllProcessors() {
-  const sharedLongTaskProcessor = LongTaskExecutor;
-  return {
-    [TransferrableMutationType.CHILD_LIST]: ChildListProcessor,
-    [TransferrableMutationType.ATTRIBUTES]: AttributeProcessor,
-    [TransferrableMutationType.CHARACTER_DATA]: CharacterDataProcessor,
-    [TransferrableMutationType.PROPERTIES]: PropertyProcessor,
-    [TransferrableMutationType.EVENT_SUBSCRIPTION]: EventSubscriptionProcessor,
-    [TransferrableMutationType.GET_BOUNDING_CLIENT_RECT]: BoundingClientRectProcessor,
-    [TransferrableMutationType.LONG_TASK_START]: sharedLongTaskProcessor,
-    [TransferrableMutationType.LONG_TASK_END]: sharedLongTaskProcessor,
-    [TransferrableMutationType.OFFSCREEN_CANVAS_INSTANCE]: OffscreenCanvasProcessor,
-    [TransferrableMutationType.OBJECT_MUTATION]: ObjectMutationProcessor,
-    [TransferrableMutationType.OBJECT_CREATION]: ObjectCreationProcessor,
-    [TransferrableMutationType.IMAGE_BITMAP_INSTANCE]: ImageBitmapProcessor,
-    [TransferrableMutationType.STORAGE]: StorageProcessor,
-    [TransferrableMutationType.FUNCTION_CALL]: FunctionProcessor,
-  };
 }
